+++ contracts/interfaces/IEpochs.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.9;
 /* SPDX-License-Identifier: UNLICENSED */
 
 interface IEpochs {
-    function getCurrentEpoch() external view returns (uint32);
+    function getCurrentEpoch() external view returns (uint256);
 
     function isStarted() external view returns (bool);
